requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
pytest>=8.0.0

# LLM Providers
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.2.0

# Web search
duckduckgo-search>=6.0.0

# Web framework
fastapi>=0.115.0
uvicorn>=0.32.0

# UI
streamlit>=1.40.0
requests>=2.32.0

# Configuration
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
